+++ apps/robot-map-gui/gui/glWidget/CRobotPose.cpp
@@ -11,13 +11,13 @@
 
 #include <mrpt/opengl/stock_objects.h>
 
-CRobotPose::CRobotPose(int id) : CSetOfObjects(), m_id(id)
+CRobotPose::CRobotPose(size_t id) : CSetOfObjects(), m_id(id)
 {
 	m_currentObj = mrpt::opengl::stock_objects::CornerXYZSimple();
 	insert(m_currentObj);
 }
 
-int CRobotPose::getId() const { return m_id; }
+size_t CRobotPose::getId() const { return m_id; }
 void CRobotPose::setSelected(bool is)
 {
 	removeObject(m_currentObj);
